requirements.txt
joblib==1.4.2
numpy==2.4.0
pandas==2.3.3
Pillow==12.1.0
scikit_learn==1.8.0
streamlit==1.40.2
